WearableApi/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.9.0
drf-spectacular==0.28.0
psycopg==3.2.10
psycopg2-binary==2.9.11
python-dotenv==1.2.1
python-decouple==3.8
celery==5.5.3
redis==7.0.1
django-celery-results==2.6.0
django-celery-beat==2.8.1
sentry-sdk==2.43.0
sendgrid==6.12.5
gunicorn==23.0.0
scikit-learn==1.7.2
numpy==2.3.4
pandas==2.3.3
joblib==1.5.2
django-sslserver==0.22